Create seperate files for specs
